SaveStackWorker: do not rely on timestamp to detect loss of frames
The timestamp is provided by DCAM and is therefore not reliable.

+++ src/gui/savestackworker.h
@@ -15,7 +15,7 @@ public:
     double getTimeout() const; // ms
     void setTimeout(double value); // ms
     void setFrameCount(int32_t count);
-    size_t getFrameCount() const;
+    int32_t getFrameCount() const;
     void setOutputFileName(const QString &fname);
     void setOutputPath(const QString &value);
 
@@ -40,7 +40,7 @@ private:
     double timeout;
     QString outputFileName;
     QString outputPath;
-    size_t frameCount, readFrames;
+    int32_t frameCount, readFrames;
     OrcaFlash *orca;
 };
